+++ check-npm-package.js
@@ -21,7 +21,7 @@ function verifyPackageExists(packageName, packageVersion) {
   const packagePath = `${ImportPathHelpers.basePath}/node_modules/${packageName}`;
   const doesPackageExist = fs.existsSync(packagePath);
   if (!doesPackageExist) {
-    logger.error(colors.red.bold(`Error checking npm module: ${packageName}@${packageVersion} (required by nathantreid:css-modules): module not found. Please ensure you have installed the module; here is the command:\n meteor npm install ${packageName} --save-dev\n`));
+    logger.error(colors.red.bold(`Error checking npm module: ${packageName}@${packageVersion} (required by wolasss:css-modules): module not found. Please ensure you have installed the module; here is the command:\n meteor npm install ${packageName} --save-dev\n`));
   }
 
   return doesPackageExist;
@@ -33,12 +33,11 @@ function checkNpmVersion(name, version) {
     if (semver.satisfies(installedVersion, version)) {
       return true;
     } else {
-      logger.warn(colors.yellow.bold(`WARNING: version mismatch for ${name}; installed version is ${installedVersion}, but version ${version} is required by nathantreid:css-modules)`));
+      logger.warn(colors.yellow.bold(`WARNING: version mismatch for ${name}; installed version is ${installedVersion}, but version ${version} is required by wolasss:css-modules)`));
       return true;
     }
   } catch (e) {
-    logger.error(colors.red.bold(`Error checking package: ${name}@${version} (required by nathantreid:css-modules): ${e.message}`));
+    logger.error(colors.red.bold(`Error checking package: ${name}@${version} (required by wolasss:css-modules): ${e.message}`));
     return false;
   }
 }
-