+++ check-npm-package.tests.js
@@ -25,7 +25,7 @@ describe('checkNpmPackage', function() {
       });
 
       checkNpmPackage('fake-package@0.0.1');
-      expect(receivedErrorMessage).to.have.string('Error checking npm module: fake-package@0.0.1 (required by nathantreid:css-modules): module not found. Please ensure you have installed the module; here is the command:\n meteor npm install fake-package --save-dev\n');
+      expect(receivedErrorMessage).to.have.string('Error checking npm module: fake-package@0.0.1 (required by wolasss:css-modules): module not found');
       done();
     });
   });
@@ -40,7 +40,7 @@ describe('checkNpmPackage', function() {
     logger.test(function() {
       logger.warn.addHook(errorMessage => {
         try {
-          expect(errorMessage).to.have.string('WARNING: version mismatch for eslint; installed version is 3.3.1, but version 1.0.0 is required by nathantreid:css-modules)');
+          expect(errorMessage).to.have.string('WARNING: version mismatch for eslint; installed version');
           done();
           return false;
         } catch (e) {