+++ test-helpers/caching-compiler.js
@@ -7,9 +7,7 @@ const fs = require('fs');
 const path = require('path');
 const createHash = require('crypto').createHash;
 const assert = require('assert');
-const Future = require('fibers/future');
 const LRU = require('lru-cache');
-const async = require('async');
 
 // Base class for CachingCompiler and MultiFileCachingCompiler.
 export class CachingCompilerBase {
@@ -20,7 +18,9 @@ export class CachingCompilerBase {
   }) {
     this._compilerName = compilerName;
     this._maxParallelism = maxParallelism;
-    const envVarPrefix = 'METEOR_' + compilerName.toUpperCase() + '_CACHE_';
+    const compilerNameForEnvar = compilerName.toUpperCase()
+      .replace('/-/g', '_').replace(/[^A-Z0-9_]/g, '');
+    const envVarPrefix = 'METEOR_' + compilerNameForEnvar + '_CACHE_';
 
     const debugEnvVar = envVarPrefix + 'DEBUG';
     this._cacheDebugEnabled = !! process.env[debugEnvVar];
@@ -32,6 +32,10 @@ export class CachingCompilerBase {
 
     // For testing.
     this._callCount = 0;
+
+    // Callbacks that will be called after the linker is done processing
+    // files, after all lazy compilation has finished.
+    this._afterLinkCallbacks = [];
   }
 
   // Your subclass must override this method to define the key used to identify
@@ -120,6 +124,14 @@ export class CachingCompilerBase {
       }, 0);
   }
 
+  // Called by the compiler plugins system after all linking and lazy
+  // compilation has finished.
+  afterLink() {
+    this._afterLinkCallbacks.splice(0).forEach(callback => {
+      callback();
+    });
+  }
+
   // Borrowed from another MIT-licensed project that benjamn wrote:
   // https://github.com/reactjs/commoner/blob/235d54a12c/lib/util.js#L136-L168
   _deepHash(val) {
@@ -132,51 +144,49 @@ export class CachingCompilerBase {
     hash.update(type + '\0');
 
     switch (type) {
-      case 'object':
-        const keys = Object.keys(val);
+    case 'object':
+      const keys = Object.keys(val);
 
-        // Array keys will already be sorted.
-        if (! Array.isArray(val)) {
-          keys.sort();
+      // Array keys will already be sorted.
+      if (! Array.isArray(val)) {
+        keys.sort();
+      }
+
+      keys.forEach((key) => {
+        if (typeof val[key] === 'function') {
+          // Silently ignore nested methods, but nevertheless complain below
+          // if the root value is a function.
+          return;
         }
 
-        keys.forEach((key) => {
-          if (typeof val[key] === 'function') {
-            // Silently ignore nested methods, but nevertheless complain below
-            // if the root value is a function.
-            return;
-          }
+        hash.update(key + '\0').update(this._deepHash(val[key]));
+      });
 
-          hash.update(key + '\0').update(this._deepHash(val[key]));
-        });
+      break;
 
-        break;
+    case 'function':
+      assert.ok(false, 'cannot hash function objects');
+      break;
 
-      case 'function':
-        assert.ok(false, 'cannot hash function objects');
-        break;
-
-      default:
-        hash.update('' + val);
-        break;
+    default:
+      hash.update('' + val);
+      break;
     }
 
     return hash.digest('hex');
   }
 
-  // We want to write the file atomically. But we also don't want to block
-  // processing on the file write.
-  _writeFileAsync(filename, contents) {
+  // Write the file atomically.
+  _writeFile(filename, contents) {
     const tempFilename = filename + '.tmp.' + Random.id();
-    fs.writeFile(tempFilename, contents, (err) => {
+
+    try {
+      fs.writeFileSync(tempFilename, contents);
+      fs.renameSync(tempFilename, filename);
+    } catch (e) {
       // ignore errors, it's just a cache
-      if (err) {
-        return;
-      }
-      fs.rename(tempFilename, filename, (err) => {
-        // ignore this error too.
-      });
-    });
+      this._cacheDebug(e);
+    }
   }
 
   // Helper function. Returns the body of the file as a string, or null if it
@@ -247,7 +257,8 @@ export class CachingCompiler extends CachingCompilerBase {
 
     // Maps from a hashed cache key to a compileResult.
     this._cache = new LRU({
-      max: this._cacheSize,
+      max: this._cacheSize || 1024 * 1024 * 10,
+      maxSize: 5000,
       length: (value) => this.compileResultSize(value),
     });
   }
@@ -275,13 +286,16 @@ export class CachingCompiler extends CachingCompilerBase {
   // you have processing you want to perform at the beginning or end of a
   // processing phase, you may want to override this method and call the
   // superclass implementation from within your method.
-  processFilesForTarget(inputFiles) {
+  async processFilesForTarget(inputFiles) {
     const cacheMisses = [];
+    const arches = this._cacheDebugEnabled && Object.create(null);
 
-    const future = new Future;
-    async.eachLimit(inputFiles, this._maxParallelism, (inputFile, cb) => {
-      let error = null;
-      try {
+    inputFiles.forEach(inputFile => {
+      if (arches) {
+        arches[inputFile.getArch()] = 1;
+      }
+
+      const getResult = () => {
         const cacheKey = this._deepHash(this.getCacheKey(inputFile));
         let compileResult = this._cache.get(cacheKey);
 
@@ -294,7 +308,7 @@ export class CachingCompiler extends CachingCompilerBase {
 
         if (! compileResult) {
           cacheMisses.push(inputFile.getDisplayPath());
-          compileResult = this.compileOneFile(inputFile);
+          compileResult = Promise.await(this.compileOneFile(inputFile));
 
           if (! compileResult) {
             // compileOneFile should have called inputFile.error.
@@ -307,19 +321,34 @@ export class CachingCompiler extends CachingCompilerBase {
           this._writeCacheAsync(cacheKey, compileResult);
         }
 
-        this.addCompileResult(inputFile, compileResult);
-      } catch (e) {
-        error = e;
-      } finally {
-        cb(error);
+        return compileResult;
+      };
+
+      if (this.compileOneFileLater &&
+          inputFile.supportsLazyCompilation) {
+        this.compileOneFileLater(inputFile, getResult);
+      } else {
+        const result = getResult();
+        if (result) {
+          this.addCompileResult(inputFile, result);
+        }
       }
-    }, future.resolver());
-    future.wait();
+    });
 
     if (this._cacheDebugEnabled) {
-      cacheMisses.sort();
-      this._cacheDebug(
-        `Ran (#${ ++this._callCount }) on: ${ JSON.stringify(cacheMisses) }`);
+      this._afterLinkCallbacks.push(() => {
+        cacheMisses.sort();
+
+        this._cacheDebug(
+          `Ran (#${
+            ++this._callCount
+          }) on: ${
+            JSON.stringify(cacheMisses)
+          } ${
+            JSON.stringify(Object.keys(arches).sort())
+          }`
+        );
+      });
     }
   }
 
@@ -350,7 +379,7 @@ export class CachingCompiler extends CachingCompilerBase {
       return;
     const cacheFilename = this._cacheFilename(cacheKey);
     const cacheContents = this.stringifyCompileResult(compileResult);
-    this._writeFileAsync(cacheFilename, cacheContents);
+    this._writeFile(cacheFilename, cacheContents);
   }
 
   // Returns null if the file does not exist or can't be parsed; otherwise