+++ test-helpers/global-variables.stub.js
@@ -7,14 +7,3 @@ import Future from 'fibers/future';
 global.Npm = {
   require
 };
-
-/* eslint-disable no-extend-native */
-Promise.prototype.await = function await() {
-  const future = new Future();
-  this.then(function(result) {
-    future.return(result);
-  }).catch(function(err) {
-    future.throw(err);
-  });
-  return future.wait();
-};